requirements.txt
# Core dependencies
grpcio>=1.50.0
grpcio-tools>=1.50.0
zeroconf>=0.60.0
protobuf>=3.20.0

# ML/Model loading
torch>=2.0.0
transformers>=4.30.0

# ONNX
onnx>=1.12.0
onnxruntime-gpu>=1.14.0

# Async
aiofiles>=23.0.0

# Optional dev dependencies
pytest>=7.0.0
black>=22.0.0
pylint>=2.15.0
